requirements.txt
streamlit>=1.32,<2.0
pandas>=2.0,<3.0
python-dotenv>=1.0,<2.0
reportlab>=4.0,<5.0
google-generativeai>=0.7.0